+++ 008.py
@@ -4,7 +4,7 @@ O salário de 6000 com o reajuste de 60% será de : 9600
 '''
 print('********** Calculadora de bônus salarial 60% **********\n')
 
-salario = int(input('Digite o valor do salário atual: '))
+salario = float(input('Digite o valor do salário atual: '))
 
 salario_ajustado = salario + (salario * 0.6)
